+++ src/ai-assistant/src/tools/geo-tools.tsx
@@ -2,7 +2,7 @@
 // Copyright contributors to the kepler.gl project
 
 import React, {useEffect} from 'react';
-import {useDispatch} from 'react-redux';
+import {useDispatch, useSelector} from 'react-redux';
 import {
   dataClassify,
   DataClassifyTool,
@@ -17,13 +17,27 @@ import {
   spatialJoin,
   SpatialJoinTool,
   spatialFilter,
-  SpatialJoinToolComponent,
   buffer,
+  BufferTool,
   centroid,
+  CentroidTool,
   dissolve,
+  DissolveTool,
   length,
   area,
-  perimeter
+  perimeter,
+  grid,
+  GridTool,
+  standardizeVariable,
+  StandardizeVariableTool,
+  thiessenPolygons,
+  ThiessenPolygonsTool,
+  minimumSpanningTree,
+  MinimumSpanningTreeTool,
+  cartogram,
+  CartogramTool,
+  rate,
+  RateTool
 } from '@openassistant/geoda';
 import {
   getUsStateGeojson,
@@ -33,18 +47,27 @@ import {
   geocoding,
   routing,
   isochrone,
-  getCachedData,
   IsochroneTool,
-  RoutingTool
+  RoutingTool,
+  roads,
+  GetUsStateGeojsonTool,
+  GetUsCountyGeojsonTool,
+  GetUsZipcodeGeojsonTool
 } from '@openassistant/osm';
+import {
+  SpatialJoinComponent,
+  DataTableComponent,
+  DataTableComponentProps
+} from '@openassistant/tables';
+import {ToolCache} from '@openassistant/utils';
 import {Datasets} from '@kepler.gl/table';
 import {Layer} from '@kepler.gl/layers';
-import {processFileData} from '@kepler.gl/processors';
 import {addDataToMap} from '@kepler.gl/actions';
 
 import {LisaToolComponent} from './lisa-tool';
-import {getGeometriesFromDataset, getValuesFromDataset} from './utils';
+import {appendColumnsToDataset, getGeometriesFromDataset, getValuesFromDataset} from './utils';
 import {AiAssistantState} from '../reducers';
+import {State} from '../components/ai-assistant-manager';
 
 export function getGeoTools(
   aiAssistant: AiAssistantState,
@@ -52,6 +75,9 @@ export function getGeoTools(
   layers: Layer[],
   layerData: any[]
 ) {
+  // tool cache
+  const toolCache = ToolCache.getInstance();
+
   // context for geo tools
   const getValues = async (datasetName: string, variableName: string) => {
     const values = getValuesFromDataset(datasets, layers, datasetName, variableName);
@@ -60,11 +86,14 @@ export function getGeoTools(
 
   const getGeometries = async (datasetName: string) => {
     let geoms = getGeometriesFromDataset(datasets, layers, layerData, datasetName);
+
     if (geoms.length === 0) {
-      // get the geoms from the cache
-      const geojson = getCachedData(datasetName);
-      if (geojson && 'features' in geojson) {
-        geoms = geojson.features;
+      // even though the tool dataset should be saved by 'saveDataToMapTool',
+      // we still try to get the dataset from tool cache, e.g. route, isochrone etc.
+      const dataset = toolCache.getDataset(datasetName);
+      // check if dataset is a GeoJSON object
+      if (dataset && dataset.type === 'geojson') {
+        geoms = dataset.content.features;
       }
     }
     return geoms;
@@ -77,82 +106,155 @@ export function getGeoTools(
     throw new Error('Mapbox token is not provided');
   };
 
+  // onToolCompleted
+  const onToolCompleted = (toolName: string, result: unknown) => {
+    toolCache.addDataset(toolName, result);
+  };
+
   // geo tools
   const classifyTool: DataClassifyTool = {
     ...dataClassify,
-    context: {
-      ...dataClassify.context,
-      getValues
-    }
+    context: {getValues}
   };
 
   const weightsTool: SpatialWeightsTool = {
     ...spatialWeights,
-    context: {
-      ...spatialWeights.context,
-      getGeometries
-    }
+    context: {getGeometries}
   };
 
   const globalMoranTool: GlobalMoranTool = {
     ...globalMoran,
-    context: {
-      ...globalMoran.context,
-      getValues
-    }
+    context: {getValues}
   };
 
   const regressionTool: SpatialRegressionTool = {
     ...spatialRegression,
-    context: {
-      ...spatialRegression.context,
-      getValues
-    }
+    context: {getValues}
   };
 
   const lisaTool: LisaTool = {
     ...lisa,
-    context: {
-      ...lisa.context,
-      getValues
-    },
+    context: {getValues},
     component: LisaToolComponent
   };
 
   const spatialJoinTool: SpatialJoinTool = {
     ...spatialJoin,
-    context: {
-      ...spatialJoin.context,
-      getValues,
-      getGeometries
-    },
-    component: CustomSpatialJoinToolComponent
+    context: {getValues, getGeometries},
+    onToolCompleted,
+    component: SpatialJoinComponent
   };
 
   const spatialFilterTool = {
     ...spatialFilter,
-    context: {
-      ...spatialFilter.context,
-      getValues,
-      getGeometries
-    },
-    component: CustomSpatialJoinToolComponent
+    context: {getValues, getGeometries},
+    onToolCompleted,
+    component: SpatialJoinComponent
   };
 
   const routingTool: RoutingTool = {
     ...routing,
-    context: {
-      ...routing.context,
-      getMapboxToken
-    }
+    context: {getMapboxToken},
+    onToolCompleted
   };
 
   const isochroneTool: IsochroneTool = {
     ...isochrone,
+    context: {getMapboxToken},
+    onToolCompleted
+  };
+
+  const bufferTool: BufferTool = {
+    ...buffer,
+    context: {getGeometries},
+    onToolCompleted
+  };
+
+  const centroidTool: CentroidTool = {
+    ...centroid,
+    context: {getGeometries},
+    onToolCompleted
+  };
+
+  const dissolveTool: DissolveTool = {
+    ...dissolve,
     context: {
-      ...isochrone.context,
-      getMapboxToken
-    }
+      getGeometries,
+      getValues
+    },
+    onToolCompleted
+  };
+
+  const roadsTool = {
+    ...roads,
+    context: {getGeometries},
+    onToolCompleted
+  };
+
+  const lengthTool = {
+    ...length,
+    context: {getGeometries}
+  };
+
+  const areaTool = {
+    ...area,
+    context: {getGeometries}
+  };
+
+  const perimeterTool = {
+    ...perimeter,
+    context: {getGeometries}
+  };
+
+  const thiessenPolygonsTool: ThiessenPolygonsTool = {
+    ...thiessenPolygons,
+    context: {getGeometries},
+    onToolCompleted
+  };
+
+  const minimumSpanningTreeTool: MinimumSpanningTreeTool = {
+    ...minimumSpanningTree,
+    context: {getGeometries},
+    onToolCompleted
+  };
+
+  const cartogramTool: CartogramTool = {
+    ...cartogram,
+    context: {getGeometries, getValues},
+    onToolCompleted
+  };
+
+  const gridTool: GridTool = {
+    ...grid,
+    context: {getGeometries},
+    onToolCompleted
+  };
+
+  const getUsStateTool: GetUsStateGeojsonTool = {
+    ...getUsStateGeojson,
+    onToolCompleted
+  };
+
+  const getUsCountyTool: GetUsCountyGeojsonTool = {
+    ...getUsCountyGeojson,
+    onToolCompleted
+  };
+
+  const getUsZipcodeTool: GetUsZipcodeGeojsonTool = {
+    ...getUsZipcodeGeojson,
+    onToolCompleted
+  };
+
+  const standardizeVariableTool: StandardizeVariableTool = {
+    ...standardizeVariable,
+    context: {getValues},
+    component: CustomDataTableComponent
+  };
+
+  const rateTool: RateTool = {
+    ...rate,
+    context: {getValues},
+    component: CustomDataTableComponent
   };
 
   return {
@@ -163,48 +265,76 @@ export function getGeoTools(
     lisaTool,
     spatialJoinTool,
     spatialFilterTool,
-    buffer,
-    centroid,
-    dissolve,
-    length,
-    area,
-    perimeter,
-    getUsStateGeojson,
-    getUsCountyGeojson,
-    getUsZipcodeGeojson,
+    gridTool,
+    bufferTool,
+    centroidTool,
+    dissolveTool,
+    lengthTool,
+    areaTool,
+    perimeterTool,
+    getUsStateTool,
+    getUsCountyTool,
+    getUsZipcodeTool,
     queryUSZipcodes,
     geocoding,
     routing: routingTool,
-    isochrone: isochroneTool
+    isochrone: isochroneTool,
+    roads: roadsTool,
+    standardizeVariable: standardizeVariableTool,
+    thiessenPolygons: thiessenPolygonsTool,
+    minimumSpanningTree: minimumSpanningTreeTool,
+    cartogram: cartogramTool,
+    rate: rateTool
   };
 }
 
 /**
- * Use SpatialJoinToolComponent to add the join result to kepler.gl
+ * Customize the DataTableComponent for StandardizeVariableTool and RateTool
  */
-function CustomSpatialJoinToolComponent(props) {
+function CustomDataTableComponent(props: DataTableComponentProps) {
   const dispatch = useDispatch();
+  const datasets = useSelector((state: State) => state.demo.keplerGl.map.visState.datasets);
+  const layers = useSelector((state: State) => state.demo.keplerGl.map.visState.layers);
+  const {originalDatasetName, datasetName, saveData} = props;
+
+  // get data by datasetName
+  const dataset = props[datasetName] as {
+    type: string;
+    content: Record<string, number[]>;
+  };
+
+  const data = dataset?.content;
 
   useEffect(() => {
-    async function addJoinedDatasetToMap() {
-      if (props.joinedDatasetId && props.joinedDataset) {
-        // add the joined dataset to kepler.gl
-        const parsedData = await processFileData({
-          content: {
-            data: props.joinedDataset,
-            fileName: props.joinedDatasetId
-          },
-          fileCache: []
-        });
-
-        dispatch(
-          addDataToMap({datasets: parsedData, options: {autoCreateLayers: true, centerMap: false}})
-        );
+    async function saveStandardizedData() {
+      // convert column-wise data to a row-wise Record<string, number>[]
+      const dataRecord: Record<string, number>[] = [];
+      const columnNames = Object.keys(data);
+      const numberOfRows = data[columnNames[0]].length;
+      for (let i = 0; i < numberOfRows; i++) {
+        const row: Record<string, number> = {};
+        for (const key of columnNames) {
+          row[key] = data[key][i];
+        }
+        dataRecord.push(row);
       }
+
+      const processedData = await appendColumnsToDataset(
+        datasets,
+        layers,
+        originalDatasetName,
+        dataRecord,
+        datasetName
+      );
+      dispatch(
+        addDataToMap({datasets: processedData, options: {autoCreateLayers: true, centerMap: false}})
+      );
     }
-    addJoinedDatasetToMap();
+    if (saveData) {
+      saveStandardizedData();
+    }
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
-  return <SpatialJoinToolComponent {...props} />;
+  return <DataTableComponent {...props} />;
 }