+++ src/ai-assistant/src/tools/kepler-tools/layer-creation-tool.tsx
@@ -6,16 +6,16 @@ import {LayerClasses} from '@kepler.gl/layers';
 import KeplerTable, {Datasets} from '@kepler.gl/table';
 import {findDefaultLayer} from '@kepler.gl/reducers';
 import {addLayer as addLayerAction} from '@kepler.gl/actions';
-import {interpolateColor} from '../utils';
-import {tool} from '@openassistant/utils';
+import {extendedTool, generateId} from '@openassistant/utils';
 import {z} from 'zod';
 import {useEffect} from 'react';
 
-export const addLayer = tool<
+export const addLayer = extendedTool<
   // parameters
   z.ZodObject<{
     datasetName: z.ZodString;
-    fieldName: z.ZodString;
+    latitudeColumn: z.ZodOptional<z.ZodString>;
+    longitudeColumn: z.ZodOptional<z.ZodString>;
     layerType: z.ZodEnum<
       [
         'point',
@@ -31,8 +31,16 @@ export const addLayer = tool<
         's2'
       ]
     >;
-    colorScale: z.ZodOptional<z.ZodEnum<['quantile', 'quantize', 'ordinal', 'custom']>>;
-    customColorScale: z.ZodOptional<z.ZodArray<z.ZodNumber>>;
+    colorBy: z.ZodOptional<z.ZodString>;
+    colorType: z.ZodOptional<z.ZodEnum<['breaks', 'unique']>>;
+    colorMap: z.ZodOptional<
+      z.ZodArray<
+        z.ZodObject<{
+          value: z.ZodUnion<[z.ZodString, z.ZodNumber, z.ZodNull]>;
+          color: z.ZodString;
+        }>
+      >
+    >;
   }>,
   // return type
   ExecuteAddLayerResult['llmResult'],
@@ -41,12 +49,36 @@ export const addLayer = tool<
   // context
   AddLayerFunctionContext
 >({
-  description: 'add a kepler.gl map layer',
+  description: `Add a kepler.gl map layer from a dataset.
+You can create basic map layer without color styling, or enhanced map layer with color visualization.
+
+For basic maps:
+- Simply use datasetName, geometryColumn (if needed), latitudeColumn/longitudeColumn (for point maps), and mapType
+- Omit color-related parameters for simple visualization
+
+For colored maps:
+- If user requests color visualization, use available columns in the dataset
+- Use dataClassify tool to classify data into bins or unique values when needed
+- If dataClassify tool returns a list of k breaks
+  a. For a list of k break values, you must create k+1 entries in the colorMap, with the last value being null.
+  b. For example: for breaks = [0, 3, 10], the colorMap could be [{value: 0, color: '##fff7bc', label: '< 0'}, {value: 3, color: '#fec44f', label: '[0-3)'}, {value: null, color: '#d95f0e', label: '>= 3'}]
+- If dataClassify tool returns a list of k unique values
+  a. There should be k colors in the colorMap. For example: for uniqueValues = ['a', 'b', 'c'], the colorMap could be [{value: 'a', color: '#1b9e77'}, {value: 'b', color: '#d95f02'}, {value: 'c', color: '#7570b3'}]
+- Generate colorBrewer colors automatically if user doesn't specify colors
+
+For geojson datasets:
+- Use geometryColumn: '_geojson' and mapType: 'geojson' even for point collections
+`,
   parameters: z.object({
     datasetName: z
       .string()
       .describe('The name of the dataset. Note: please do NOT use the datasetId.'),
-    fieldName: z.string(),
+    latitudeColumn: z.string().optional(),
+    longitudeColumn: z.string().optional(),
+    layerName: z
+      .string()
+      .optional()
+      .describe('If possible, generate a name for the layer based on the context.'),
     layerType: z.enum([
       'point',
       'arc',
@@ -60,13 +92,16 @@ export const addLayer = tool<
       'trip',
       's2'
     ]),
-    colorScale: z.enum(['quantile', 'quantize', 'ordinal', 'custom']).optional(),
-    customColorScale: z
-      .array(z.number())
-      .optional()
-      .describe(
-        'An array of numeric breakpoints used to define custom color intervals. Only applicable when colorScale is set to "custom"'
+    colorBy: z.string().optional(),
+    colorType: z.enum(['breaks', 'unique']).optional(),
+    colorMap: z
+      .array(
+        z.object({
+          value: z.union([z.string(), z.number(), z.null()]),
+          color: z.string()
+        })
       )
+      .optional()
   }),
   execute: executeAddLayer,
   component: AddLayerToolComponent
@@ -76,20 +111,17 @@ export type AddLayerTool = typeof addLayer;
 
 type AddLayerArgs = {
   datasetName: string;
+  layerName?: string;
   layerType: string;
-  fieldName: string;
-  colorScale: string;
-  customColorScale: Array<number>;
+  latitudeColumn?: string;
+  longitudeColumn?: string;
+  colorBy?: string;
+  colorType?: string;
+  colorMap?: Array<{value: string | number | null; color: string; label?: string}>;
 };
 
 function isAddLayerArgs(args: any): args is AddLayerArgs {
-  return (
-    typeof args === 'object' &&
-    args !== null &&
-    'datasetName' in args &&
-    'fieldName' in args &&
-    'layerType' in args
-  );
+  return typeof args === 'object' && args !== null && 'datasetName' in args && 'layerType' in args;
 }
 
 type AddLayerFunctionContext = {
@@ -124,7 +156,16 @@ async function executeAddLayer(args, options): Promise<ExecuteAddLayerResult> {
       throw new Error('Invalid addLayer context');
     }
 
-    const {datasetName, fieldName, layerType, colorScale = 'quantile', customColorScale} = args;
+    const {
+      datasetName,
+      layerName,
+      latitudeColumn,
+      longitudeColumn,
+      layerType,
+      colorBy,
+      colorType,
+      colorMap
+    } = args;
 
     const datasets = options.context.getDatasets();
 
@@ -136,76 +177,98 @@ async function executeAddLayer(args, options): Promise<ExecuteAddLayerResult> {
 
     // check if field exists in the dataset
     const dataset = datasets[datasetId];
-    const field = dataset.fields.find(f => f.name === fieldName);
-    if (!field) {
-      throw new Error(`Field ${fieldName} not found.`);
-    }
-
-    // check colorScale is valid
-    if (!['quantile', 'quantize', 'ordinal', 'custom'].includes(colorScale)) {
-      throw new Error(`Invalid color scale: ${colorScale}.`);
-    }
-
-    // check if customColorScale is available
-    if (colorScale === 'custom' && !customColorScale) {
-      throw new Error('Custom color scale or breaks is required when colorScale is "custom".');
-    }
 
     // check if layerType is valid
-    const layer = guessDefaultLayer(dataset, layerType);
+    let layer = guessDefaultLayer(dataset, layerType);
 
+    const layerId = layer?.id || `layer_${generateId()}`;
+
+    if (!layer) {
+      // for point layer, try to creat a point layer manually if LLM sugggests Lat/Lng fields
+      if (layerType === 'point' && latitudeColumn && longitudeColumn) {
+        layer = {
+          id: layerId,
+          type: 'point',
+          config: {
+            dataId: datasetId,
+            label: layerName || `${datasetName}-${layerType}`,
+            columns: {
+              lat: {value: latitudeColumn, fieldIdx: dataset.getColumnFieldIdx(latitudeColumn)},
+              lng: {value: longitudeColumn, fieldIdx: dataset.getColumnFieldIdx(longitudeColumn)}
+            }
+          },
+          visConfig: {
+            colorRange: {
+              name: 'Ice And Fire',
+              type: 'diverging',
+              category: 'Uber',
+              colors: ['#D50255', '#FEAD54', '#FEEDB1', '#E8FEB5', '#49E3CE', '#0198BD']
+            }
+          }
+        } as any;
+      }
+    }
     if (!layer) {
       throw new Error(`Invalid layer type: ${layerType}.`);
     }
 
-    const colorField = {
-      name: field.name,
-      type: field.type
-    };
-
-    // create custom colorRange if needed
-    let customColorRange = layer.config.visConfig.colorRange;
-    if (colorScale === 'custom') {
-      const newColors = interpolateColor(customColorRange.colors, customColorScale.length + 1);
-      customColorRange = {
-        ...customColorRange,
-        colors: newColors,
-        colorMap: newColors.map((color, index) => [customColorScale[index] || null, color])
-      };
-    }
+    const columns = layer?.config?.columns || {};
 
     // construct new layer config for addLayer() action
     const newLayer = {
-      id: layer.id,
+      id: layerId,
       type: layer.type,
       config: {
         ...layer.config,
         dataId: datasetId,
-        label: `${field.name}-${colorScale}`,
-        columns: Object.keys(layer.config.columns).reduce((acc, key) => {
-          acc[key] = layer.config.columns[key].value;
+        label: layerName || `${datasetName}-${layerType}`,
+        columns: Object.keys(columns).reduce((acc, key) => {
+          const column = columns[key];
+          if (column) {
+            acc[key] = column.value;
+          }
           return acc;
-        }, {}),
-        colorScale,
-        colorField,
-        strokeColorScale: colorScale,
-        strokeColorField: colorField,
-        visConfig: {
-          ...layer.config.visConfig,
-          colorRange: customColorRange,
-          strokeColorRange: customColorRange,
-          ...(customColorScale
-            ? {colorDomain: customColorScale, strokeColorDomain: customColorScale}
-            : {})
-        }
+        }, {})
       }
     };
+
+    if (colorBy) {
+      const colorField = dataset.fields.find(f => f.name === colorBy);
+      if (!colorField) {
+        throw new Error(`Field ${colorBy} not found.`);
+      }
+      // create kepler.gl's colorMap from uniqueValues and breaks
+      const colorScale = colorType === 'breaks' ? 'custom' : 'customOrdinal';
+      const colors = colorMap?.map(color => color.color);
+      const keplerColorMap = colorMap?.map(color => [color.value, color.color]);
+      const colorRange = {
+        name: 'color.customPalette',
+        type: 'custom',
+        category: 'Custom',
+        colors,
+        colorMap: keplerColorMap
+      };
+
+      newLayer.config['colorScale'] = colorScale;
+      newLayer.config['colorField'] = colorField;
+      newLayer.config['strokeColorScale'] = colorScale;
+      newLayer.config['strokeColorField'] = colorField;
+      newLayer.config.visConfig['colorRange'] = colorRange;
+      newLayer.config.visConfig['strokeColorRange'] = colorRange;
+      newLayer.config['visualChannels'] = {
+        colorField: {
+          name: colorBy,
+          type: colorField?.type
+        },
+        colorScale
+      };
+    }
 
     return {
       llmResult: {
         success: true,
         layer: JSON.stringify(newLayer),
-        details: `map layer with ${field.name} and ${colorScale} color scale will be added to the map.`
+        details: `map layer ${layerId} will be added to the map.`
       },
       additionalData: {
         layer: newLayer,