+++ src/duckdb/src/table/duckdb-table.ts
@@ -38,11 +38,10 @@ import {
 } from './duckdb-table-utils';
 
 import {
-  constructST_asWKBQuery,
+  castDuckDBTypesForKepler,
   dropTableIfExists,
   getDuckDBColumnTypes,
   getDuckDBColumnTypesMap,
-  getGeometryColumns,
   removeUnsupportedExtensions,
   restoreArrowTable,
   restoreUnsupportedExtensions
@@ -127,6 +126,7 @@ export class KeplerGlDuckDbTable extends KeplerTable {
       await c.query(createTableSql);
     } catch (error) {
       console.log('importRowData', error);
+      throw error;
     }
   }
 
@@ -143,9 +143,11 @@ export class KeplerGlDuckDbTable extends KeplerTable {
         FROM ST_READ('${this.id}', keep_wkb = TRUE);
         ALTER TABLE '${this.label}' RENAME '${DUCKDB_WKB_COLUMN}' TO '${KEPLER_GEOM_FROM_GEOJSON_COLUMN}';
       `;
+
       await c.query(createTableSql);
     } catch (error) {
       console.error('importGeoJsonData', error);
+      throw error;
     }
 
     return {
@@ -180,6 +182,7 @@ export class KeplerGlDuckDbTable extends KeplerTable {
       // Known issues:
       // 1) Arrow Type with extension name: geoarrow.point and format: +w:2 is not currently supported in DuckDB.
       console.error('importArrowData', error);
+      throw error;
     }
 
     return {
@@ -236,8 +239,7 @@ export class KeplerGlDuckDbTable extends KeplerTable {
 
       const duckDbColumns = await getDuckDBColumnTypes(c, tableName);
       const tableDuckDBTypes = getDuckDBColumnTypesMap(duckDbColumns);
-      const columnsToConvertToWKB = getGeometryColumns(duckDbColumns);
-      const adjustedQuery = constructST_asWKBQuery(tableName, columnsToConvertToWKB);
+      const adjustedQuery = castDuckDBTypesForKepler(tableName, duckDbColumns);
       const arrowResult = await c.query(adjustedQuery);
 
       // TODO if format is an arrow table then just use the original one, instead of the new table from the query?
@@ -252,6 +254,7 @@ export class KeplerGlDuckDbTable extends KeplerTable {
         .filter(col => col) as arrow.Vector[];
     } catch (error) {
       console.error('DuckDB table: createTableAndGetArrow', error);
+      throw error;
     }
 
     await c.close();